add fragment work in progress - date picker work
to make the attachment button work ( the attachment putton is on the appbar that is menaged by the mainActivity ) i needed somethin to set the click listener dinamically.
so, i made an activity view model that is shared between all fragments and allow to set the click listners.

now i can implement the click listener of the fab in a fragment!

in the mainActivity the viewmodel is observed and change dynamically what the click listener are!